Basic file rename for consistency
All other files use snake_case names, so updated the couple that use
camelCase to snake_case for consistency.
